requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
slowapi==0.1.9
pyjwt==2.8.0
bcrypt==4.1.2
